[FEATURE] FieldValueHelp: support collective search
Add a collectiveSearchItems aggregation to FieldValueHelp.
If more than one item is assigned, a collectiveSearchSelect control is
added to the FilterBar of the value help dialog.
If the CollectiveSearch is switched the contentRequest function of the
FieldHelp delegate is called. As property the key of the selected
collectiveSearch is provided.
In suggestion mode, always the default collectionveSearch is used (first
item).

Change-Id: I908980ab00a3fb2d867a1d5705f4bdf479eaaecc
JIRA: CPOUIFDENVER-1303
JIRA: CPOUIFDENVER-1574