[INTERNAL] Opa: use already available sinon in xhr waiter
This reverts commit 26340fb56afa52691bdfe02ed89771c8a6f72ef8.

Reason for revert: the fallback to sync loading of sinon.js causes an
async/sync conflict when e.g. sap/ui/core/util/MockServer.js loads
sinon in an asynchronous context. As sinon is not an AMD module, the
conflicting second execution can'T be aborted by the loader and, when 
executed, overrides any settings that have been made. 

In the anaylzed context of the Cart demo app, the useFilters property
(set to true by the MockServer) was set back to the default false and
caused unmocked requests to fail. As the conflict depends on the timing
of the async loading, it occurred randomly.

Change-Id: I933edbedd51ab84403211f273fea90d280d92cd5