[FIX] sap.uxap.HierarchicalSelect: Fixed button's focus on all themes
Problem: Focus of the HierarchicalSelect drop-down button is incorrect.

Root Cause: Recent regression caused in the sap.m.Select, where we no
longer apply focus on the root DOM element of the Select, but rather
on one of its children (hiddenSelect).

Solution: We correct the CSS rules which expect ":focus" on the root
DOM element of the Select, to actually expect the Select to have the
"sapMFocused" CSS class on its root when it's focused.

Additionally a small regression is fixed where the items in the AnchorBar
overlap with the overflow button.

Change-Id: I5e14dc237f8996f8074c637eccd162cfc774d64e
BCP: 2070468666