[INTERNAL][FIX] sap.m.SplitContainer: Added private method _safeBackToPage
In historical reason (error was thrown when the page is not in the history stack)
the _safeBackToPage was implemented in the sap.m.NavContainer which is called by
sap.f.TargetHandler which is used in sap.f.Router.

sap.f.routing.TargetHandler has specific login for sap.m.NavContainer,
sap.m.SplitContainer and sap.f.FlexibleColumnLayout controls.

Problem:
Error is thrown when the sap.m.SplitContainer is used with sap.f.Router.

Root cause:
Error is thrown because sap.m.SplitContainer doesn't have proxy function which
points to private functions  of the internally used sap.m.NavContainer

Fix:
Added the missing proxy function.

Change-Id: I276079b01fb59ee7cf430ffdf1a9ddc94075df49