[INTERNAL] MdcTable: API to scroll the table to a given index and focus a row
scrollToIndex returns a promise that resolves when the table is scrolled.
setFocusOnRow scrolls the table if necessary and sets the focus to the first interactive element of
the row. If there are no interactive elements, sets the focus on the entire row in case of m.Table,
and on the first cell in case of ui.Table

Change-Id: If3dd6f54d11b345a92dd6baaaa9f3977e25fb60e
JIRA: CPOUIFTEAMB-1601