[INTERNAL] sap.m.MultiComboBox, sap.m.ComboBox: RTL support improvements
RTL support on a control level is implemented for sap.m.ComboBOx and sap.m.MultiComboBox.

The InputBase renderer is fixed to add a direction attribute
to the wrapper element instead of the inner one, if textDirection property is set.

The picker's textDirection in the ComboBox and MultiComboBox is synced with the
textDirection property of the controls.

BCP: 002075129400007952042020
Change-Id: I11c303787658898211fb3f928418e65bfd1c2ff9