[INTERNAL] ExtensionPoint: Pass correct View instance and Fragment ID to EP Provider
The EP Provider needs the Fragment's explicitly set ID to distinguish
between multiple occurences of the same Fragment.

Additionally we now always pass a View instance to the EP Provider.
Before this change, if the Fragment is created via the factory,
the EP Provider received a Fragment instance instead.

Change-Id: I7ebf8ba5b561c3fbb14e91ea23404c163ba2ca80
BCP: 002075129500007902632020
CR-Id: 002075125900004539272020