[INTERNAL][FIX] sap.uxap.ObjectPageHeaderActionButton: Improved accessibility
Problem:
Some accessibility attributes are not set to the DOM element.

Root cause:
Accessibility of the sap.m.Button is based on whether the text inside it
is visible. In sap.uxap.ObjectPageHeaderActionButton the text is hidden
with a CSS class and its parent (sap.m.Button) can't generate correct
accessibility states.

Fix:
Removed the CSS class that hides the text and adjusted the
sap.uxap.ObjectPageHeaderActionButton to use the hook provided by
sap.m.Button that allows the child control to control the text state.

BCP: 2070474123
Change-Id: Iff80fd2136efdd5cd12682941dd1f103a33b93d8