[INTERNAL] sap.ui.core..modifiers: replace type checks with 'isA'
the 'isA' function was not available when the type checks were
implemented, but the functionality is exactly the same. Only that the
current implementation has some shortcomings:
when a control extends from a control that implements an interface. isA
correctly detects that, the old implementation does not

Change-Id: Ia77072b2393e2ed2dd1bf894e8434b3c01365043
BCP: 2040790263
CR-Id: 002075125900004489132020
(cherry picked from commit 14929dc91f49488a0a905db599c3f7f8fc1df3be)