[FIX] sap.f.DynamicPage: HeaderExpanded redundant space fixed
Problem: in a snippix (94265) the view is being rendered twice, so this problem cannot be
reproduced in our samples, unless we trigger invalidate() on the DynamicPage after initial
rendering. The problem occurs when headerExpanded is set to 'false' and content is enough for a
scrollbar to appear. When the control is rerendered, the "_canSnapHeaderOnScroll" method is
falsly returning 'true', because the header is already marked being in the title area, and we
are passing case, which is different than the one we are dealing with.

Solution: we calculate the maxScrollPosition related to the header height only in case we have
actual scroll height in the dynamic page.
BCP: 002075129400008032622020
CR-Id: 002075125900004643262020
Change-Id: I2585545424eacb8a3f958ed10d5483768617d8d1
(cherry picked from commit 79f4174df51e7a597acde23f75a1f470d1300700)