[INTERNAL] TreeItemBase: fix double announcement when tree item is expanded
When the tree item is expanded the _updateItem function was updating the
aria-expanded=true and also this tree item rerenders. Due to the above 2
changes, the tree item which was expanded was being announced twice.
The _updateItem function was there to update the tree item structure with
relevant DOM attributes, styleClasses and tree item indentations. But with
the new DOM patch we can simple rerender the treeItem so that the correct
tree item state can be applied by the renderer.
BCP: 2080397055

Change-Id: I89aa61992cb2fd0073e9726d6f0645c50cb9b035