[INTERNAL][FIX] sap.m.OverflowToolbarButton: Improved accessibility
Problem:
Item's text is read twice.

Root cause:
Text and tooltip have same value and the screen reader reads them
both.

Fix:
Now, hook that allows to add custom logic for tooltip value is added.
Tooltip is shown when the button is outside of the overflow area (icon
only button) and when the tooltip is different to text, otherwise the tooltip is hidden.

In addition test file is added for sap.m.OverflowToolbarButton.

BCP: 2070468513
Change-Id: Ie0ad21b0858ea676eb9cf6dbb616c384ebec34a8