[INTERNAL] sap.ui.fl - mulitple extensionpoints in fragments
With fragment instantiation in controller coding it is possible to
have multiple extension points with the same key into the application.
The fragment has then an template character. When it is instantiated
multiple times the included extensionpoints have still the same name.
All controls added on extensionpoint gets now prefixed with the
fragment xml name from the outside to avoid duplicate id clashes.

Change-Id: I819adda6332c9cbca4ca398936cd9809a8fd7518
CR-Id: 002075125900004629182020
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
JIRA: TEAM42-100
BCP: 2040790263