[FIX] base.util.Storage: avoid concurrent modification while looping
The previous implementation of the 'removeAll' method assumed that the
implementation of the Storage Web API does not reorder entries when a
certain key is removed.

As we got reports that the method sometimes misses to remove some
entries, the implementation has been changed so that it no longer relies
on the above assumption. The keys-to-remove are first collected and only
then removed.

Change-Id: I1aaf7ce22e24d379a7d0bcba1ea5fc558d873b0f
BCP: 2080404229