[INTERNAL] Demo Kit: Enabled binary download in Demo Apps page
Excluded the logic for loading of the items in separated module
which now is used in SampleBaseController.js and DemoApps.controller.js

Allow list was changed to contain text file extensions that are
used in the samples, otherwise files will be treated as binary files.

DemoApps download dialog was changed to sap.m.Dialog instead of
sap.m.SelectDialog because onDownloadPress function was called twice.

JIRA: BGSOFUIPIRIN-4935
Change-Id: I6b54f73364ddf5c7e9b3dc487bfe8cedbf1ac4d5