[INTERNAL] sap.m.Select: Replace native HTML select
Problem: Native dropdown menu can't be prevented on some ios devices.

Solution: Remove the native select element and align the select implementation
with the Select-Only combobox example from aria 1.2 practices page.

Impact: Improve the functionality usage and screen reader experience

Risk: There are some known issues in regards to screen reader and aria attributes, which are adressed to screen reader owners and in process to be fixed.

Notes:
 1) Direct text node of div@role="combobox" is passed to tha Accessibility VALUE API;
 2) INPUT element is needed for the cases when sap.m.Select is part of a form and the
form is submited by button@type="submit";
 3) When chaning the selection with popover closed the new value is not announced on MacOS
and the InvisibleMessage is used
 4) Value state and value state message changes are also announced with the InvisibleMessage.
Screen Readers are announcing every change of elements with aria-live attributes. When the data
is later populated after contro initialization this lead to confusing speech output when there are a lot of
Select controls.

Change-Id: I59e30f7d558ae0d0f8a2d422408bfc7642a300ed
CR-Id: 002075125900000548562021
BCP: 2070432748
BCP: 002075129400000393512021