[FIX] mvc.XMLView: Provide better logging for processing errors
When synchronously processing an XMLView the XMLTemplateProcessor
did not report all errors correctly.

Previously we tried to throw the errors, yet we have seen that
this has caused multiple applications to break.

We now still accept the broken XML and instantiate the conrols,
but log an error for easier debugging.

Change-Id: I72b2328c198914bb7fd464ed9dd4a89d7d9353c7
JIRA: CPOUI5FRAMEWORK-202
Fixes: https://github.com/SAP/openui5/issues/2968
BCP: 2080425463
BCP: 2080321485