[FIX] sap.f.AvatarGroup: Fixed border-radius of group more button
Problem: When there is a big number in the group more button of the
AvatarGroup, its border is incorrect and it doesn't look like it's
specified in the visual spec.

Root Cause: The 'border-radius' CSS property of the button is incorrect.

Solution: We set the 'border-radius' CSS property of the button to 3.5rem
as recommended by VD.

Change-Id: I69b2d0c31755f904ca6a98c93662816f58ccc270
BCP: 2080426370