[INTERNAL][FIX] ExtensionPoint: Provide Fragment ID if not prefixed
The original change mistakenly assumed that getLocalId() on the
View class will always return an ID.
In case the given ID is not prefixed by the View ID, null is returned.

This change recognizes the situation and passes the original
expicitly given Fragment ID to the ExtensionPoint Provider.

Change-Id: I1e4177b271528c835a6a0a6a959098de831da537
BCP: n/a