[INTERNAL] Form: use always same title size independend of Level
The title level for Form and FormContainer can be changed to meet the
page structure. This should not change the visual appearence of the
titles.
For compatibility reasons this is only changed in Fiori3 themes.
In Fiori3 the FormContainer tirle is per default now also rendered as h4
title to have no gaps in hierarchy.

Change-Id: I9ec4aac1ca399afb0645c0d623b52f6a511d38c3
BCP: 2080418011
BCP: 2080410501
BCP: 2080400740
BCP: 2070501019
BCP: 2070478570
BCP: 2070480517