[FIX] sap.m.OverflowToolbar: Overlapping of buttons is fixed
Problem: When there is a SearchField in the OverflowToolbar, and there
is not enough space for all controls, some buttons overlap the
SearchField.

Solution: The SearchField has min-width specified, which does not allow
its inner content to overflow out of its main container. However, in
Toolbar.less we override this min-width for all shrinkable items in the
context of Toolbar/OverflowToolbar. As a result, the inner content of
the SearchField overflows out of its container, which leads to
overlapping of the controls, placed inside the OverflowToolbar.
As a solution, now we do not override the default min-width of the
SearchField in Toolbar.less.

Change-Id: I89a9f06a1942aa8e5570f82a6855be341ec8608c
BCP: 2080433034