[INTERNAL][FIX] sap.ui.fl - multiple extensionpoint in fragments
With fragment instantiation in controller coding it is possible to
have multiple extension points with the same key into the application.
The fragment has then an template character. When it is instantiated
multiple times the included extensionpoints have still the same name.
All controls added on extensionpoint gets now prefixed with the

Change-Id: I7e83cfd2b314907d48adabfab445586bc52c638d
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
JIRA: TEAM42-100
BCP: 2040790263