[FIX] sap.m.DateRangeSelection: can select a range in month with 4 weeks
When there are only 4 weeks in the current month, there is a dummy
DOM element placed as a placeholder of the 5th missing week.
This dummy element should not be considered as part of the month
for the calculations inside the control.

BCP: 2080437818
BCP: 002075129400000540672021
CR-Id: 002075125900000693732021
Fixes: https://github.com/SAP/openui5/issues/3104
Change-Id: I15f9967b1cb29b8d2c9aed68670a4af671ec0456
(cherry picked from commit bedbd4f4e46fff809af3d167fbb97cb880a7e63b)