[INTERNAL] sap.ui.fl: use old way to save with multiple namespaces
variants are saved in a different namespace than all other flex objects,
so when a new variant is saved together with ui changes there are
multiple namespaces. As the new condense route is not enabled to handle
that in one request the old changes route is used and only dirty changes
get condensed.
Also the SaveAs scenario does not support backend condensing for now and
 is disabled

Change-Id: I88d5ebbe6b11301e916f373e7d566a98206f2ba6
BCP: 2080443065