[INTERNAL] sap.ui.fl: fix JS applying of template changes
the change applying was blocked because of selectors pointing to the
template. Those controls don't go through our propagation listener and
the changes never got applied. Now the dependent selectors are simply
ignored in that case.
This might lead to racing conditions with lazy loading and then changes
being applied when the necessary controls are not there yet. But we
can't wait for controls inside templates

Change-Id: Idf7a704b8e51a58bf754ff106424d9c90cd4725e
BCP: 002075129400008714412020