[INTERNAL] sap.m.PlanningCalendar: correct picker shown after changing view
When the user opens the yearpicker on hours view, changes the view
to day view and opens the picker again, the year picker is shown instead of
the day picker. Now the month view of the picker is shown in the end.
There were analogical issues in all of the views.

The scenario is covered by visual tests.

Change-Id: I498c92d2310c26fb206554e8c617a3fce8037eb7