[INTERNAL] sap.ui.fl - processing extension points in template
This issue occurs just when an extension point is used into aggregation
templates and this template is defined into fragment. When an extension
point change needs to be applied in this scenario then we can use just
the js application procedure. For this we need to change the selector
to the target control id. In this case we require as selector the id
from the parent control that defines the aggregation template instead
of the aggregation templat id itself.

Change-Id: Ic28f8f17986a5e9b1e151dbdc36df4809a81578e
CR-Id: 002075125900000107222021
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
JIRA: TEAM42-100
BCP: 2040790263