[INTERNAL] sap.m.InputBase: Control-level RTL support is now reverted
This reverts commit e18548f437ef6a4caa1a4b4f6ad5328bb2055022.
The reverted change was created in result of BCP issue
002075129400007952042020 in order to reordering of the inner control
elements based of the text-direction property. For example - when
it is set to RTL the tokens to be aligned to the right and ordered
from right to left.

However this change introduced many visual bugs and unexpected results
and should be reverted as it was before - the textDirection property is now
again only affecting the text value within the native input element insider of
the control.

BCP: 2180003960
Change-Id: I11c303787658898211fb3f928418e65bfd1c1cfa