[FEATURE] sap.m.Image: Lazy loading attribute added
Now, sap.m.Image has the option to use the 'loading' attribute of
HTML img element with value 'lazy' to make the control less heavy
on the apps.
The default behaviour of the cntrol is not affected with the change.
The change takes effect only after the 'lazyLoading' property is set
to 'true'.

JIRA: BGSOFUIPIRIN-4938
Change-Id: Ib9ad7fcd1af4b403930b9d5db5ec3757caee3444