[FIX] sap.f.DynamicPage: Improved sample
Problem:
If we modify the sample via devTools to show the page footer,
the analytical table (nested inside the DynamicPage) starts
flickering at a certain page height. This is a documented limitation of the analytical table.

Solution:
As discussed with the table owners, we need to workaround the table limitation
from our (DynamicPage) side, by enabling the fitContent property.

BCP: 2070413091
Change-Id: I0fc8c27810a59daf63f025941004035fcea0f71d