[INTERNAL] ListBase: prevent meta key handling by the ItemNavigation
The "alt/meta + left/right" keyboard shortcut is used by the browser to
navigate backward and forward in the browser history.
"alt" in Windows and "meta" in macOS.
The ListBase disabled only the "alt" key for the ItemNavigation. Now also
adding the "meta" key to the disabledModifiers of the ItemNavigation, since
navigating backward and forward was prevented in macOS.
BCP: 2070196219

Change-Id: Ie41f3c3faf4065449259aaa648e04f107a7a6352