[INTERNAL] REUSE improvements for license tracking of mixed files
Maintain license&copyright information for files with multiple
licenses/copyrights no longer within the source file, but also in the
dep5 file.
To enable this, enrich the .license file information for these "partial"
files with information about the other libraries which contribute to the
same source file(s).
Also, add the version number of the generator used to the dep5 file, so
a consistency check knows with which version of the generator the dep5
file has been created.

Change-Id: I3fd188b9b33cc1a60172a16b5f821b2157de7c42