[FIX] sap.m.OverflowToolbar: "Type" Button property now invalidates Button
[Problem] When sap.m.Button is inside sap.m.OverflowToolbar and its "type" property is changed
to "Emphasized", the font-weight of the Button is set to bold, which changes the width of the
button. The Overflow Toolbar doesn`t recalculate the content size, because this property of the
button is part of its configuration as part of OverflowToolbarAssociativePopoverControls and
is in the "noInvalidationProps" list.

[Solution] We removed that property form the no invalidation list, since its require re-
calculation of content width.

BCP: 2080412801
CR-Id: 002075125900000536802021
Change-Id: Ie422c64386a9ee06b158a821a884cd85783dd8d3
(cherry picked from commit 4a8605bf6a5dc36993479704a09ca2c3155a6f0b)