[INTERNAL] Table: fix autoPopinMode recalculation
The logic to prevent the invalidation of the table and recalculation for
the
autoPopinMode caused some unexpected behavior in the table.
The sap.m.Column control is already modified not to call the table
renderer
often when columns get invalidated. This will aid in performing better
calculation for every change that is done on the column instance.
This also makes some old logic unnecessary which can simply be removed
and
the Table can rerender and perform recalculation for the autoPopinMode
accordingly.
BCP: 2180005247

Change-Id: Ifc1721aebaed2728f008b843e545d3713bf63fb2