[FIX] sap.m.MultiInput: Failing tests in IE are now fixed
The qUnit tests are now updated according to the new input alignment
changes in sap.m.Tokenizer. Now the click on the nMore indicator
has the same behaviour on IE as in the other browsers. There is no
need for the removed 'if' statemets and the alternating testing for
MS IE anymore.

CR-Id: 002075125900000293892021
BCP: 2180012233
Change-Id: Ia2a44ea4fe52b12f150b4d7e3c8d23b3ee7afac3