[INTERNAL] core.XMLTemplateProcessor: fully implement semantic rendering
Also modifies assertions in RenderManager to allow rendering CSS styles
either by calling style("prop",...) or attr("style",...), but not both.
So far only style() was allowed. Similarly, either class("class1") or
attr("class", "...") can be used, but not both.

Allowing to write the style and class attributes as a whole simplifies
the parsing of embedded (x)HTML in XMLViews.

Change-Id: I6346430f58e7a10925441f1ea0657fff337a05ef