[INTERNAL] GrowingEnablement: fix _getHasScrollbars function
The GrowinEanblement._getHasScrollbars returned true if the rendered
items are greater than or equal to 40. Due to this the growing trigger
button is hidden in UI when 40 item are rendered. But in large screen
resolution, it is possible that there are no scrollbars rendered so the
user is not able to load the next page after 40 items are loaded. This
check is now adapted, where we check the number of items that can be fit
on the screen and the growing trigger button is rendered accordingly
BCP: 2180019925

Change-Id: Ia0bea36149e9089d9d7076ab4100dd80125b5268