[INTERNAL] FieldHelp: use operator of FilterField
If the FilterField has no equal(EQ) operator but a similar custom
operator, the
FieldHelp needs to use this operator for selecting items.

Also on parsing direct user input the custom operator needs to be used
for validation.

On stringify to save in variants such custom operators needs to be
supported.

Change-Id: Id7787f2c7f0ed837b813d9bf2550a9f451122ffc
JIRA: CPOUIFDENVER-1576