[INTERNAL] sap.m.ListItemBase: LESS calculation removal
The sap.m.Tree background color needs to be changed to @sapUiListAlternatingBackground for all level BG except Level 1. in order to remove the less calculation.
In addition, after consulting with the Fiori design colleagues, also the calculation for the alternative row color should be removed and use the default instead.

Jira: CPOUIFTEAMB-1540
Change-Id: Id137e7fed06e952567bc64a01ef4fb84c273e3c7