[INTERNAL] P13nConditionPanel: No assertions are shown for missing resoureces of custom types
Problem:
When a type is specified for Include/Exclude operations, this type is used
to get the translations from the resource boundle. In that case if the resources
are missing an assertion is shown in the console.

Solution:
Add a check if type is present and if the resource is available in the boundle
before trying to get it, otherwise fallback to resource without a type.

BCP: 2180032308
Change-Id: Ifcba38c0fc8f621a871466baa5eee9aa2adef57d