[FIX] sap.m.DateTimePicker: popover can now be reopened
Problem description:
If the control popover was opened and after that closed
via mouse click outside of its DOM representation,
then it couldn't be reopened again.

Solution:
Time in miliseconds for opening and closing animations
of the popup is added, so the opening and closing
event handlers are properly ordered in the event queue.

Change-Id: I8002af96ea57e5865164dd16c32dbbe794aa0262
CR-Id: 002075125800000885292022
BCP: 2180043591
BCP: 2280025389