[INTERNAL] sap.ui.core: let marker interfaces extend class Control
Using 'extends' for interfaces additionally requires a fix in the SDK as
it so far makes the implicit assumption that borrowed members can only
exist when the current entity also has ui5 metadata. For interfaces,
this is not the case. The fix in ApiDetail.controller.js just avoids a
runtime error, but the resulting display might not match expectations.
When looking at an interface that requires sap.ui.core.Control as base
class of implementations, one most likely is not interested in seeing
the methods etc. of sap.ui.core.Control, too. This UX question needs to
be discussed and addressed separately.

Change-Id: Ic29c1ce2f9e38f308a2b6bad7d81221bd1a6eb5f