[FEATURE] Localization: add Serbian (Latin), enhance Locale mappings
CLDR data:
----------
In addition to the existing sr.json (which contains CLDR data for
Serbian (Cyrillic)), a new file sr_Latn.json is generated, containing
data for Serbian (Latin).

ISO Code Mappings:
------------------
The mapping from the old ISO639 language code "in" to the new code "id"
and vice versa has been removed as it was never used for messagebundles.
With this change, only the new code "id" is supported.

SAP Logon Language:
-------------------
The mapping of SAP Logon languages to BCP47 language tags and vice versa
has been enhanced with the following mappings:

  BCP47  | SAP Logon Language
  -------+-------------------
  en-GB  |   6N
  fr-CA  |   3F
  es-MX  |   1X
  pt-PT  |   1P

The mappings apply when the sap-language URL parameter is given (its
value will be converted from SAP Logon Language to BCP47) or when one of
the Locale#getSAPLogonLanguage() or Configuration#getSAPLogonLanguage()
methods is called (Locale will be converted to a SAP Logon Language).

Accept-Language Header:
-----------------------
Method Configuration#getLanguageTag, which is mainly used by model
implementations to fill the 'Accept-Language' header, maps sr-Latn to
sh in order to remain backward compatible.

Others:
-------
A new method LocaleData#getCurrentLanguageName() is introduced, which
returns a display name for the current language.

BCP: 1780409720
JIRA: CPOUI5FOUNDATION-306
Change-Id: I7e13d1cdfc2fc5221ad1e0556843dc846c8d3ba7