[FIX] sap.f.DynamicPage: Infinite snapping/unsnapping is prevented
Problem: When DynamicPage is resized, in some cases, the Header snaps
and unsnaps infinitly. This happens because every time the Header moves
from content to title part and vice versa, DynamicPage calls its method
for overriding the preserveHeaderStateOnScroll property and this again
snap/unsnaps the Header.

Soluiton: When onChildControlsHeight function is called, we check if the
change in the Header's height is due to its movement from content to
title part or vice versa, and do not call the method for overriding the
preserveHeaderStateOnScroll function, as it has already been called.
Second call leads to snapping/unsnapping which is not needed and leads
to infinite operations.

Change-Id: I32a1ef59c9b46b1f9c98c227cd9727ea5504d539
BCP: 0020751294 0000040909 2021