[INTERNAL] sap.uxap.ObjectPageLayout: Updated color of scroll container
Problem: When there is a container with a custom background-color under
the ObjectPageLayout and the header is expanded but no spacer is
needed, the background color of the section may differ with the one
below it (of the application). This results in different applications
adjusting their app background-color in order to match the one of the
section, which is inconvenient.

Solution: We apply the background color of the section on the whole
scroll wrapper of the ObjectPageLayout. This way there is no option for
color miss-match between the section background and the background of the
application.

Change-Id: I10e7e252e699dabe7d4d68c6a7ec37af4b3c9573
BCP: 2080466267