[FIX] sap.f.AvatarGroup: Focuses in HCB and HCW aligned with spec
Problem: Focuses of the AvatarGroup in "group" mode, each Avatar in
"individual" mode and the overflow button are not aligned with the
visual specification for HCB and HCW theme. Currently their width is
0.0625 rem, but it should be 0.125rem.

Solution: Override focus width for HCB and HCW theme to 0.125rem, as per
the visual specification.

Change-Id: Ie658735f3fa24e93d5ac743e475652462a421b2e
BCP: 2180068783