[FIX] sap.m.Breadcrumbs: Fix overflowing of items when not needed
Problem: When navigating to another page and the breadcrumbs get out of
view, after returning on the previous page, some items of the
breadcrumbs are overflowed without the need to.

Solution: When breadcrumbs get out of view, they are hidden and the
resize handler is triggered with current width 0. In this case, we do
not need to update anything and calculate which items should be
shown/hidden, as currently the whole breadcrumbs are hidden. Now we
prevent this wrong calculations and as a result - when getting back to
the previous page, breadcrumbs occupy again the needed space and
calculations are done correctly.

Change-Id: I23936b493f1c4bb68c29e12e0fb505eedc7a6228
BCP: 2180046291