[FIX] sap.m.PlanningCalendar: the view key is always set after all the views are available
Issue: if the view key is set before all of the custom views
are available, some of the inner controls are not initialised
and not all of the inner settings of the control are set properly.

Solution: ensure that the view key is set after the corresponding
view is available.

BCP: 2180080181
Change-Id: If5c1fa3396cf51cda544ebb5fac5e62d63860f6e