[FIX] sap.f.DynamicPage: Workaround table flickering
Problem:
When the DynamicPage renders both a footer and a sap.ui.table.Table,
and the page height is under a certain threashold, the table starts
infinite flickering.
This is a documented limitation of the sap.ui.table.Table that we hit when
DynamicPage adds a footer spacer. The table docu contains: "The table must
be rendered without siblings in its parent DOM element" but in our case the
footer-spacer is a sibling.

Solution:
In the above scenario (when the page is scrollable) replace the footer-spacer
with a padding-bottom.

Note:
In the alternative case, when fitContent is enabled, replaced the former
spacer with ::after element, to ensure the footer does not overlap the page
content.

Change-Id: I040dd43b1e1430afbdc914b86f1a068fdb673820
BCP: 2070413091
JIRA: BGSOFUIPIRIN-5025