[INTERNAL] Element.enhanceAccessibilityState: fix implementations / docs
The enhanceAccessibilityState method was mistakenly documented as
returning the enhanced state object. But RenderManager rather expects it
to modify the second parameter "mAriaProps" (by reference, so to say).

The unnecessary return statement has been removed from implementations,
documentation has been adapted and tests that relied on the return value
have been corrected (as they tested a behavior that the RenderManager
didn't care about).

Change-Id: Id74e88986984bde644e248f1a3ce186c2d6df5e7