[INTERNAL][FIX] Table: Requesting contexts for fixed bottom rows
If the binding is sorted/filtered/refreshed/..., the length might be 0 (V2 and
V4 behave differently). To avoid that the table requests incorrect indices for
fixed bottom rows that might lead to additional and unnecessary data requests
(or even errors - V4), the cached binding length is ignored when getting
contexts for fixed bottom rows. Instead, the current binding length is used.

Change-Id: I1c83a57e19548dce30a65741918a617c84dafec2