[FIX] sap.ui.model.analytics.AnalyticalBinding: support selecting only the text of a measure
The analytical binding must not fail if only the text property of a
measure has been added a column but not the measure itself. The behavior
shall be the same as if the measure with its total has been requested.

Test:
http://localhost:8080/testsuite/resources/sap/ui/test/starter/Test.qunit.html?testsuite=test-resources/sap/ui/core/qunit/testsuites/testsuite.databinding.qunit&test=analytics/AnalyticalBinding

PS1: red test
PS2: issue fixed
PS3: fixed similar code places (no tests)
PS4: that instead of this
PS5: added missing tests
PS6-8: review comments

Change-Id: Ie7195a4ae07c90cf1eef08228a80de09f5f81cce
BCP: 002075129500000644142021