[FIX] DemoKit: Wait for correct SWA initialization event
Problem:
Demokit waits for the loading of the SWA tracker script before
starting it with swa.enable.
However, 'load' event of the script is not sufficient to ensure
the tracker is initializaed prior to calling swa.enable

Solution:
Subscribe for the "swaLoadSuccess" event fired by the SWA tracker
after its initialization.

Change-Id: Id5726d49fe95487f45944c8f4adaf99a9561487b
BCP: 2180085944