[INTERNAL]Engine: provide validation mechanism for p13n UIs
Engine supports UI validation which can be triggered in the
BaseController derivations.

The according BaseController derivation needs to implement "model2State".
As the validation is service depending, a delegate hook "validateState"
has been introduced which is being executed synchronously every time a
change on the UI has been done.

AggregationBaseDelegate#validateState will be called with the complete
Control state, while the BaseController derivations will calculate how
the state "would" look like if the user would press "OK" (model2State result).

The method returns an object defining the message type and the text itself,
which is going to be displayed in the according p13n UI.

Jira: CPOUIFDENVER-1597
Change-Id: I12661fbde7f09b0b05719f8df0db1e2891b29329