[FIX] CalendarRow: Appointments with no overlap are rendered in the same level
Problem:
Two appointments - the first ends where the second starts,
are rendered one above the other as if their times overrlap.
This happens due to arbitrary rounding of their start dates.
Solution:
No rounding is applied to the appointments' start dates.
So appointments which does not overlap, are rendered
in the same level.

Change-Id: I2f4f1e7ccdec5f1077e004780d20ed85aeba19ac
BCP: 002075129400000342992021