[INTERNAL] Demo Kit: Added appearance change feature
There is a new menu in the "More Information" button menu.
It is called "Appearance". It changes the appearance of the Demo Kit
to "Light" (bright theme), "Dark" (dark theme) or "Auto" (depending on
OS settings).
The user's choice is saved in a cookie, if the cookies are enabled.

Additionally, the "More Information" button's icon is changed.
Also there are icons for all of the options in its menu.

Some the block layout tiles (In Home and Documentation pages) are
changed in order to appeal on different appearances.

Those changes are synced with VD and are mentioned in the backlog.

Change-Id: I48aace96dc8a15bad8f3426c5fe38044cd64faae
JIRA: BGSOFUIPIRIN-4946