[INTERNAL] DemoKit: New Temporary banner added
With this change we add ability to show important message banner in Demo Kit.
It can be closed, so that it won`t appear anymore in the browser. The banner also includes
date of expiration, which will be limiting the time range of its appearance.

With this change we are adding a message to inform about the end of IE11 support.

Jira: BGSOFUIPIRIN-5039
Change-Id: I0c3a723a3f22e0efe3c90f569998ca2d5b018a4a