[INTERNAL] fetchProperties/ExtensionsForBinding
Introduction of 2 new methods in mdc.odata.V4.TableDelegate :
fetchPropertiesForBinding and fetchPropertyExtensionsForBinding, allowing
to have different sets of propertyInfos used in the table and in the
binding.
For example, this allows to have properties that are part of the
analytical query but not shown as aggregatable/groupable in table
column headers.

PS2: fixed trailing commas in QUnit test file
PS3: moved the flattening of visible propertyInfos back in the delegate
PS4: destroy transient propertyHelper

Jira: CPOUIFTEAMB-1762
Change-Id: I60bea4f2673c99fa45404b43b8c10310b238919c