[INTERNAL] ColumnResizer: fix resize handle circle creation
In iPad Pro and iPad Air devices the sap.ui.Device API returns unexpected
output where desktop="true" for these devices. Due to this the resize handle
circle was not created by the plugin.
To work around this issue we now make use of the window.matchMedia to detect
if mouse pointers are available on the device.
BCP: 2180114977

Change-Id: Id61adb1d8098b1ec7eef18dfc910e0f5fcb1f195