[INTERNAL] sap.m.Input: Suggestion rows/columns observer
This change introduces suggestions table observer. The observer
allows us to act on changes to the table's items and columns
aggregations.
This helps us in two ways:
- We remove repetitive code;
- We do not overwrite the methods for inserting / removing
suggestion rows / columns in the Input. Those aggregations
are directly forwarded to the suggestions table, so it is
more appropriate to observe it and react on changes in its
aggregations, than overwriting methods inherited from the
ManagedObject.

Change-Id: I5d1e7d3f4fd0d5a23d13a04da3277639a0b25af8
JIRA: BGSOFUIRILA-3047