[INTERNAL] Demo Kit: Correct short description in chuncked topics
Background: The element with class "shortdesc" is used as a subtitle to the
ObjectPage. To not be displayed again in the html ObjectPage content the element
is removed.

Problem: When there are multiple or nested topicis only the first occurrence
of the element should be remove.

Solution: Introduce new array to hold selectors whose elements can be queried
using querySelector instead of querySelectorAll

Change-Id: Ia81fea9f2b14c5aee334730cfbbc53b0994e55ab
BCP: 2180107504