[INTERNAL][FIX] AnalyticalTable: Group indentation after expand
Problem description:

After updating the binding contexts of rows after an expand, the indentation for
the group structure is applied. The AnalyticalTable changes the default
indentation, which was designed for the experimental grouping feature of the
base table. Chrome is not able to handle such dynamic and synchronous changes
to styles and does not paint the indentation correctly. The styles are correctly
set in the DOM, though.

Solution:

The issue is solved by implementing the group indentation of the AnalyticalTable
as the default.
For the group mode in the TreeTable, the leaf rows are not aligned with their
group header.
The experimental grouping overwrites the default.

Change-Id: I4d12cda5453434123a9df70f413c435b3c1be93f
BCP: 2180042196
CR-Id: 002075125900001306332021