[INTERNAL][FIX] AnalyticalTable: Group indentation after expand
Problem description:

After updating the binding contexts of rows after an expand, the indentation for
the group structure is applied. The AnalyticalTable changes the default
indentation, which was designed for the experimental grouping feature of the
base table. Chrome is not able to handle such dynamic and synchronous changes
to styles and does not paint the indentation correctly. The styles are correctly
set in the DOM, though.

Solution:

The issue is solved by implementing the group indentation of the AnalyticalTable
as the default. This also solves the issue that the indentation is different in
the AnalyticalTable and the base table with the V4Aggregation plugin.

For the group mode in the TreeTable, leaf rows are not aligned with their group
headers.

The experimental grouping feature does not overwrite the default, hence the
indentation is changed now: There simply is no indentation anymore. It supports
only one level of grouping, and leaf rows are at the same level as their group
headers.
Please note that in the downports, the experimental grouping overwrites the
default to avoid visual regressions in already released versions. This is not
done in the master branch for consistency, and to be really sure to avoid the
Chrome bug in all cases.

Change-Id: I4d12cda5453434123a9df70f413c435b3c1be93f
BCP: 2180042196